Own the ChdbError the engine raises
The chdb wrapper ships a chdb/__init__.py that shadows chdb-core's and
defines its own ChdbError(Exception), so raising chdb.ChdbError handed
callers that class instead. chdb's datastore falls back from a streaming
query to a synchronous one by catching RuntimeError, which stopped
catching it, and its test suite failed 64 of the tests in the files this
touches. Keep the class in chdb/_exceptions.py, which the wrapper does
not ship, so the engine raises one RuntimeError-derived error either way.
