+++ chdb/state/sqlitelike.py
@@ -4,8 +4,8 @@ from typing import Optional, Any
 import sys
 from decimal import Decimal
 from urllib.parse import parse_qsl
-import chdb
 from chdb import _chdb
+from chdb._exceptions import ChdbError
 from chdb.progress_display import (
     get_notebook_display as _get_notebook_display,
     is_notebook as _is_notebook,
@@ -179,7 +179,7 @@ class StreamingResult:
         except Exception as e:
             self._exhausted = True
             self._cleanup_progress_callback_once()
-            raise chdb.ChdbError(f"Streaming query failed: {str(e)}") from e.with_traceback(None)
+            raise ChdbError(f"Streaming query failed: {str(e)}") from e.with_traceback(None)
 
     def __iter__(self):
         return self
@@ -252,7 +252,7 @@ class StreamingResult:
             try:
                 self._conn.streaming_cancel_query(self._result)
             except Exception as e:
-                raise chdb.ChdbError(f"Failed to cancel streaming query: {str(e)}") from e.with_traceback(None)
+                raise ChdbError(f"Failed to cancel streaming query: {str(e)}") from e.with_traceback(None)
             finally:
                 self._cleanup_progress_callback_once()
         else:
@@ -755,7 +755,7 @@ class Connection:
             else:
                 result = self._conn.query(query, format, params=params or {})
         except RuntimeError as e:
-            raise chdb.ChdbError(str(e)) from e.with_traceback(None)
+            raise ChdbError(str(e)) from e.with_traceback(None)
         else:
             return result_func(result)
         finally:
@@ -866,7 +866,7 @@ class Connection:
             c_stream_result = self._conn.send_query(query, format, params=params or {})
         except RuntimeError as e:
             self._cleanup_auto_progress_callback(progress_callback)
-            raise chdb.ChdbError(str(e)) from e.with_traceback(None)
+            raise ChdbError(str(e)) from e.with_traceback(None)
         except Exception:
             self._cleanup_auto_progress_callback(progress_callback)
             raise
@@ -1092,7 +1092,7 @@ class Cursor:
         self._cursor.execute(query)
         result_mv = self._cursor.get_memview()
         if self._cursor.has_error():
-            raise chdb.ChdbError(self._cursor.error_message())
+            raise ChdbError(self._cursor.error_message())
         if self._cursor.data_size() == 0:
             self._current_table = None
             self._current_row = 0
